chore: move vector2 file into include/global
